Use long because json integers can exceed the maximum allowed value for avro int

+++ src/index.js
@@ -5,7 +5,7 @@ const typeMapping = {
   string: 'string',
   null: 'null',
   boolean: 'boolean',
-  integer: 'int',
+  integer: 'long',
   number: 'float',
 }
